using correct data type of unit32_t for colour values
the previous type of int is only 16bit and cannot hold the 24bit of colour info. The effect is that the most significant colour (red here) is trimmed off, and can't display any red. Same for the param type of displayNumber(). 

Also we are changing the comms type from RGB to GRB...we suspect that this previously a hack to get around above loss of color bits. Clear the clockFace and Downlight comms mode should be the same. They came from the same strip of LEDs

Need second commit to change param types in digit functions